Rename class to represent PluginVersion
