+++ securesignals-ima/src/main/java/com/uid2/securesignals/ima/PluginVersion.kt
@@ -1,21 +1,16 @@
-package com.uid2.securesignals.gma
+package com.uid2.securesignals.ima
 
 import com.uid2.BuildConfig
 import com.uid2.Version
 
 /**
- * An object exposing the version information associated with the UID2 GMA Plugin.
+ * An object exposing the version information associated with the UID2 IMA Plugin.
  */
-object UID2SecureSignals {
+internal object PluginVersion {
     private const val VERSION_STRING = BuildConfig.VERSION
 
     private const val VERSION_COMPONENTS = 3
     private val INVALID_VERSION = Version(0, 0, 0)
-
-    /**
-     * Gets the version of the included UID2/IMA plugin, in string format.
-     */
-    fun getVersion() = VERSION_STRING
 
     /**
      * Gets the version of the included UID2/IMA plugin, in its individual major, minor and patch components.
